requirements.txt
# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Background Tasks
celery==5.3.4
redis==5.0.1

# AI & ML
openai==1.3.7
langchain==0.0.350
langchain-openai==0.0.2
pinecone-client==2.2.4

# File Processing
pypdf2==3.0.1
python-docx==1.1.0
pandas==2.1.3
unstructured==0.10.30
tiktoken==0.5.1

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Email
resend==0.6.0

# Calendar Integration
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Payment Processing
stripe==7.8.0

# File Storage
boto3==1.34.0

# Utilities
pydantic==2.5.0
pydantic-settings==2.1.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
